Fixed compilation error with Java 1.6 in ClientRingbufferProxy.

+++ hazelcast-client/src/main/java/com/hazelcast/client/proxy/ClientRingbufferProxy.java
@@ -56,7 +56,7 @@ public class ClientRingbufferProxy<E> extends ClientProxy implements Ringbuffer<
     public long capacity() {
         if (capacity == -1) {
             CapacityRequest request = new CapacityRequest(getName());
-            capacity = invoke(request);
+            capacity = (Long) invoke(request);
         }
 
         return capacity;
